pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "ember"
version = "0.1.0"
description = "ML for Core Electron Heating due to Wave Modulation in the Solar Wind"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cdflib",
    "matplotlib",
    "numpy",
    "pandas",
    "scipy",
    "seaborn",
]

[project.optional-dependencies]
anomaly = [
    "PyWavelets",
    "scikit-learn",
    "torch",
    "torchvision",
    "zuko",
]
anomaly-notebook = [
    "PyWavelets",
    "lightgbm",
    "open_clip_torch",
    "scikit-learn",
    "timm",
    "torch",
    "torchvision",
    "umap-learn",
    "zuko",
]
dev = [
    "pre-commit",
    "pytest",
]
test = [
    "pytest",
]

[project.scripts]
ember-download = "ember.cli:main_download"
ember-spectrograms-dvac = "ember.cli:main_spectrograms_dvac"
ember-spectrograms-vac = "ember.cli:main_spectrograms_vac"
ember-anomaly = "ember.cli:main_anomaly"

[tool.setuptools.packages.find]
where = ["."]
include = ["ember*"]
